requirements.txt
Pillow==10.3.0
ultralytics==8.2.2
opencv-python==4.8.1.78
torch==2.1.2
